+++ Thesis-Fight/Assets/Scripts/Stats/Stat.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class Stat
 {
 
-    private int _finalValue;
+    private float _finalValue;
 
     // Properties
     public string StatName { get; set; }
     public string StatDescription { get; set; }
-    public int BaseValue { get; set; }
-    public int FinalValue
+    public float BaseValue { get; set; }
+    public float FinalValue
     {
         get
         {
@@ -27,7 +27,7 @@ public class Stat
     public List<IStatModifier> Debuffs;
 
     // Constructors
-    public Stat(string _statName, string _statDescription, int _baseValue)
+    public Stat(string _statName, string _statDescription, float _baseValue)
     {
         StatName = _statName;
         StatDescription = _statDescription;
